Time: 128 ms (45.78%) | Memory: 112.8 MB (17.17%) - LeetSync

+++ 643-maximum-average-subarray-i/maximum-average-subarray-i.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     double findMaxAverage(vector<int>& nums, int k) {
-        int window=0;
+        double window=0;
         double mxSum =0;
 
         for(int i=0;i<k;i++){
@@ -11,7 +11,7 @@ public:
 
         for(int i=k;i<nums.size();i++){
             window += nums[i] - nums[i-k];
-            mxSum = max(mxSum , (double)window);  // MT
+            mxSum = max(mxSum , window);  // MT
             
         };
         return mxSum/k;
